requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-seed==0.1.9
djangorestframework==3.10.2
Faker==2.0.0
mysqlclient==1.4.4
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
